// client/web/src/schema/site.schema.d.ts
export interface SiteConfiguration {
// ...
  htmlBodyTop?: string;
// ...
  "auth.providers"?: (
    | BuiltinAuthProvider
    | SAMLAuthProvider
    | OpenIDConnectAuthProvider
    | HTTPHeaderAuthProvider
    | GitHubAuthProvider
    | GitLabAuthProvider
  )[];
// ...
}
// ...
export interface BuiltinAuthProvider {
  type: "builtin";
  /**
   * Allows new visitors to sign up for accounts. The sign-up page will be enabled and accessible to all visitors.
   *
   * SECURITY: If the site has no users (i.e., during initial setup), it will always allow the first user to sign up and become site admin **without any approval** (first user to sign up becomes the admin).
   */
  allowSignup?: boolean;
}
/**
 * Configures the SAML authentication provider for SSO.
 *
 * Note: if you are using IdP-initiated login, you must have *at most one* SAMLAuthProvider in the `auth.providers` array.
 */
export interface SAMLAuthProvider {
  type: "saml";
  /**
   * An identifier that can be used to reference this authentication provider in other parts of the config. For example, in configuration for a code host, you may want to designate this authentication provider as the identity provider for the code host.
   */
  configID?: string;
  /**
   * The name to use when displaying this authentication provider in the UI. Defaults to an auto-generated name with the type of authentication provider and other relevant identifiers (such as a hostname).
   */
  displayName?: string;
  /**
   * The SAML Service Provider name, used to identify this Service Provider. This is required if the "externalURL" field is not set (as the SAML metadata endpoint is computed as "<externalURL>.auth/saml/metadata"), or when using multiple SAML authentication providers.
   */
  serviceProviderIssuer?: string;
  /**
   * The SAML Identity Provider metadata URL (for dynamic configuration of the SAML Service Provider).
   */
  identityProviderMetadataURL?: string;
  /**
   * The SAML Identity Provider metadata XML contents (for static configuration of the SAML Service Provider). The value of this field should be an XML document whose root element is `<EntityDescriptor>` or `<EntityDescriptors>`. To escape the value into a JSON string, you may want to use a tool like https://json-escape-text.now.sh.
   */
  identityProviderMetadata?: string;
  /**
   * The SAML Service Provider certificate in X.509 encoding (begins with "-----BEGIN CERTIFICATE-----"). This certificate is used by the Identity Provider to validate the Service Provider's AuthnRequests and LogoutRequests. It corresponds to the Service Provider's private key (`serviceProviderPrivateKey`). To escape the value into a JSON string, you may want to use a tool like https://json-escape-text.now.sh.
   */
  serviceProviderCertificate?: string;
  /**
   * The SAML Service Provider private key in PKCS#8 encoding (begins with "-----BEGIN PRIVATE KEY-----"). This private key is used to sign AuthnRequests and LogoutRequests. It corresponds to the Service Provider's certificate (`serviceProviderCertificate`). To escape the value into a JSON string, you may want to use a tool like https://json-escape-text.now.sh.
   */
  serviceProviderPrivateKey?: string;
  /**
   * The SAML NameID format to use when performing user authentication.
   */
  nameIDFormat?: string;
  /**
   * Sign AuthnRequests and LogoutRequests sent to the Identity Provider using the Service Provider's private key (`serviceProviderPrivateKey`). It defaults to true if the `serviceProviderPrivateKey` and `serviceProviderCertificate` are set, and false otherwise.
   */
  signRequests?: boolean;
  /**
   * Whether the Service Provider should (insecurely) accept assertions from the Identity Provider without a valid signature.
   */
  insecureSkipAssertionSignatureValidation?: boolean;
  /**
   * Allows new visitors to sign up for accounts via SAML authentication. If false, users signing in via SAML must have an existing Sourcegraph account, which will be linked to their SAML identity after sign-in.
   */
  allowSignup?: boolean;
}
/**
 * Configures the OpenID Connect authentication provider for SSO.
 */
export interface OpenIDConnectAuthProvider {
  type: "openidconnect";
  /**
   * The name to use when displaying this authentication provider in the UI. Defaults to an auto-generated name with the type of authentication provider and other relevant identifiers (such as a hostname).
   */
  displayName?: string;
  /**
   * An identifier that can be used to reference this authentication provider in other parts of the config. For example, in configuration for a code host, you may want to designate this authentication provider as the identity provider for the code host.
   */
  configID?: string;
  /**
   * The URL of the OpenID Connect issuer.
   *
   * For Google Apps: https://accounts.google.com
   */
  issuer: string;
  /**
   * The client ID for the OpenID Connect client for this site.
   *
   * For Google Apps: obtain this value from the API console (https://console.developers.google.com), as described at https://developers.google.com/identity/protocols/OpenIDConnect#getcredentials
   */
  clientID: string;
  /**
   * The client secret for the OpenID Connect client for this site.
   *
   * For Google Apps: obtain this value from the API console (https://console.developers.google.com), as described at https://developers.google.com/identity/protocols/OpenIDConnect#getcredentials
   */
  clientSecret: string;
  /**
   * Only allow users to authenticate if their email domain is equal to this value (example: mycompany.com). Do not include a leading "@". If not set, all users on this OpenID Connect provider can authenticate to Sourcegraph.
   */
  requireEmailDomain?: string;
}
/**
 * Configures the HTTP header authentication provider (which authenticates users by consulting an HTTP request header set by an authentication proxy such as https://github.com/bitly/oauth2_proxy).
 */
export interface HTTPHeaderAuthProvider {
  type: "http-header";
  /**
   * The name (case-insensitive) of an HTTP header whose value is taken to be the username of the client requesting the page. Set this value when using an HTTP proxy that authenticates requests, and you don't want the extra configurability of the other authentication methods.
   */
  usernameHeader: string;
  /**
   * The prefix that precedes the username portion of the HTTP header specified in `usernameHeader`. If specified, the prefix will be stripped from the header value and the remainder will be used as the username. For example, if using Google Identity-Aware Proxy (IAP) with Google Sign-In, set this value to `accounts.google.com:`.
   */
  stripUsernameHeaderPrefix?: string;
  /**
   * The name (case-insensitive) of an HTTP header whose value is taken to be the email of the client requesting the page. Set this value when using an HTTP proxy that authenticates requests, and you don't want the extra configurability of the other authentication methods.
   */
  emailHeader?: string;
}
// ...
export interface GitHubAuthProvider {
  type: "github";
  /**
   * URL of the GitHub instance, such as https://github.com or https://github-enterprise.example.com.
   */
  url?: string;
  /**
   * The Client ID of the GitHub OAuth app, accessible from https://github.com/settings/developers (or the same path on GitHub Enterprise).
   */
  clientID: string;
  /**
   * The Client Secret of the GitHub OAuth app, accessible from https://github.com/settings/developers (or the same path on GitHub Enterprise).
   */
  clientSecret: string;
  /**
   * The name to use when displaying this authentication provider in the UI. Defaults to an auto-generated name with the type of authentication provider and other relevant identifiers (such as a hostname).
   */
  displayName?: string;
  /**
   * Allows new visitors to sign up for accounts via GitHub authentication. If false, users signing in via GitHub must have an existing Sourcegraph account, which will be linked to their GitHub identity after sign-in.
   */
  allowSignup?: boolean;
  /**
   * Restricts new logins to members of these GitHub organizations. Existing sessions won't be invalidated. Leave empty or unset for no org restrictions.
   */
  allowOrgs?: string[];
  /**
   * Experimental: Allows sync of GitHub teams and organizations permissions across all external services associated with this provider to allow enabling of [repository permissions caching](https://docs.sourcegraph.com/admin/repo/permissions#permissions-caching).
   */
  allowGroupsPermissionsSync?: boolean;
}
// ...
export interface GitLabAuthProvider {
  type: "gitlab";
  /**
   * URL of the GitLab instance, such as https://gitlab.com or https://gitlab.example.com.
   */
  url?: string;
  /**
   * The Client ID of the GitLab OAuth app, accessible from https://gitlab.com/oauth/applications (or the same path on your private GitLab instance).
   */
  clientID: string;
  /**
   * The Client Secret of the GitLab OAuth app, accessible from https://gitlab.com/oauth/applications (or the same path on your private GitLab instance).
   */
  clientSecret: string;
  /**
   * The name to use when displaying this authentication provider in the UI. Defaults to an auto-generated name with the type of authentication provider and other relevant identifiers (such as a hostname).
   */
  displayName?: string;
  /**
   * The OAuth API scope that should be used
   */
  apiScope?: "api" | "read_api";
}